simplify flask implementation
