+++ .gitignore
@@ -48,3 +48,4 @@ logs/
 test_files/
 *.json
 *.har
+__pycache__/
